Use __vector unsigned long in PPC test programs

+++ TestPrograms/test_ppc_power7.cxx
@@ -24,8 +24,8 @@ int main(int argc, char* argv[])
     __vector unsigned int x = {5,6,7,8};
     __vector unsigned int y = vec_xl(0, (unsigned int*)argv[0]);
     __vector unsigned int z = vec_xor(x, y);
-    __vector unsigned long long xx = {1,2};
-    __vector unsigned long long yy = (__vector unsigned long long)y;
+    __vector unsigned long xx = {1,2};
+    __vector unsigned long yy = (__vector unsigned long)y;
 #else
     int x[-1];
 #endif
